UPDATED README, NEW GAME Using Graphics Engine API. Well structured in a folder, I present the newly packaged PYTHONTERMINALGRAPHICSAPI. I am pleased to say this is the newst version of the engine.
We are using the most advanced features presently written to create multiple layers and scenes in order to give the user experience the most positive feedback.

I have used maps, map readers, classes and functions to create a more dynamic and interactive experience.
New use case of code is the ability to add animations and interact with certain sections of the screen without interfering with the rest of the drawing onto screen.

With well mapped out code, I have created an efficient and poweful way to display graphics using the terminal.